[proliferate] final frontend session mechanical cleanup
