+++ desktop/src/components/workspace/shell/tabs/ChatTabsMenu.tsx
@@ -8,7 +8,7 @@ import {
 import { createPortal } from "react-dom";
 import { PopoverMenuItem } from "@/components/ui/PopoverMenuItem";
 import { ChevronRight } from "@/components/ui/icons";
-import { recordSubagentChildRelationshipHint } from "@/hooks/sessions/session-relationship-hints";
+import { recordSubagentChildRelationshipHint } from "@/hooks/sessions/workflows/session-relationship-hints";
 import type { HeaderChatMenuEntry } from "@/hooks/workspaces/tabs/workspace-header-tabs-view-model-types";
 import type { HeaderSubagentChildRow } from "@/hooks/workspaces/tabs/use-workspace-header-subagent-hierarchy";
