+++ desktop/src/hooks/chat/use-chat-prompt-actions.ts
@@ -6,12 +6,12 @@ import {
 import { useWorkspaceSetupStatusCache } from "@/hooks/access/anyharness/workspaces/use-workspace-setup-status-cache";
 import { useSessionCreationActions } from "@/hooks/sessions/use-session-creation-actions";
 import { useSessionRuntimeActions } from "@/hooks/sessions/use-session-runtime-actions";
-import { useSessionPromptWorkflow } from "@/hooks/sessions/use-session-prompt-workflow";
+import { useSessionPromptWorkflow } from "@/hooks/sessions/workflows/use-session-prompt-workflow";
 import { useSessionCancelActions } from "@/hooks/sessions/workflows/use-session-cancel-actions";
 import { useSessionFindOrCreateActions } from "@/hooks/sessions/workflows/use-session-find-or-create-actions";
 import { isSessionModelAvailabilityInterruption } from "@/hooks/sessions/workflows/use-session-model-availability-workflow";
 import { useSessionPromptActions } from "@/hooks/sessions/workflows/use-session-prompt-actions";
-import { useSessionSelectionActions } from "@/hooks/sessions/use-session-selection-actions";
+import { useSessionSelectionActions } from "@/hooks/sessions/facade/use-session-selection-actions";
 import { useChatInputStore } from "@/stores/chat/chat-input-store";
 import { useToastStore } from "@/stores/toast/toast-store";
 import { useSessionSelectionStore } from "@/stores/sessions/session-selection-store";
