+++ desktop/src/hooks/sessions/facade/use-session-selection-actions.ts
