+++ desktop/src/hooks/sessions/lifecycle/session-runtime-helpers.ts
