+++ desktop/src/hooks/sessions/lifecycle/session-runtime-pending-config.ts
