+++ desktop/src/hooks/sessions/workflows/session-activation-guard.ts
