+++ desktop/src/hooks/sessions/workflows/session-control-contract.ts
@@ -2,7 +2,7 @@ import type {
   ContentPart,
   PromptInputBlock,
 } from "@anyharness/sdk";
-import type { SessionActivationGuard, SessionActivationOutcome } from "@/hooks/sessions/session-activation-guard";
+import type { SessionActivationGuard, SessionActivationOutcome } from "@/hooks/sessions/workflows/session-activation-guard";
 import type { PromptAttachmentSnapshot } from "@/lib/domain/chat/composer/prompt-attachment-snapshot";
 import type { MeasurementOperationId } from "@/lib/domain/telemetry/debug-measurement-catalog";
