+++ desktop/src/hooks/sessions/workflows/session-creation-in-flight.ts
