+++ desktop/src/hooks/sessions/workflows/session-creation-request-options.ts
@@ -1,0 +1,6 @@
+import { getLatencyFlowRequestHeaders } from "@/lib/infra/measurement/latency-flow";
+
+export function buildLatencyRequestOptions(latencyFlowId?: string | null) {
+  const headers = getLatencyFlowRequestHeaders(latencyFlowId);
+  return headers ? { headers } : undefined;
+}
