+++ desktop/src/hooks/sessions/workflows/session-mode-preferences.ts
