+++ desktop/src/hooks/sessions/workflows/use-session-config-actions.ts
@@ -1,7 +1,7 @@
 import { useSetSessionConfigOptionMutation } from "@anyharness/sdk-react";
 import { useCallback } from "react";
 import { useWorkspaceSessionCache } from "@/hooks/access/anyharness/sessions/use-workspace-session-cache";
-import { persistDefaultSessionModePreference } from "@/hooks/sessions/session-mode-preferences";
+import { persistDefaultSessionModePreference } from "@/hooks/sessions/workflows/session-mode-preferences";
 import type { SessionConfigOptionUpdateOptions } from "@/hooks/sessions/workflows/session-control-contract";
 import { useWorkspaceRuntimeBlock } from "@/hooks/workspaces/use-workspace-runtime-block";
 import { useWorkspaceSurfaceLookup } from "@/hooks/workspaces/use-workspace-surface-lookup";
