+++ desktop/src/hooks/sessions/workflows/use-session-prompt-actions.ts
@@ -1,5 +1,5 @@
 import { useCallback } from "react";
-import { useSessionPromptWorkflow } from "@/hooks/sessions/use-session-prompt-workflow";
+import { useSessionPromptWorkflow } from "@/hooks/sessions/workflows/use-session-prompt-workflow";
 import type { ActiveSessionPromptOptions } from "@/hooks/sessions/workflows/session-control-contract";
 import { useWorkspaceRuntimeBlock } from "@/hooks/workspaces/use-workspace-runtime-block";
 import {
