+++ desktop/src/hooks/sessions/workflows/use-session-prompt-workflow.ts
