+++ desktop/src/hooks/sessions/workflows/use-session-selection-actions.test.ts
