+++ desktop/src/hooks/sessions/workflows/use-session-summary-actions.ts
@@ -17,8 +17,8 @@ import {
 import { buildSessionSlotPatchFromSummary } from "@/lib/domain/sessions/summary";
 import { activityFromTranscript } from "@/lib/domain/sessions/directory/directory-activity";
 import { batchSessionStoreWrites } from "@/lib/infra/scheduling/react-batching";
-import { persistDefaultSessionModePreference } from "@/hooks/sessions/session-mode-preferences";
-import { clearPendingConfigRollbackCheck } from "@/hooks/sessions/session-runtime-pending-config";
+import { persistDefaultSessionModePreference } from "@/hooks/sessions/workflows/session-mode-preferences";
+import { clearPendingConfigRollbackCheck } from "@/hooks/sessions/lifecycle/session-runtime-pending-config";
 import { useWorkspaceSurfaceLookup } from "@/hooks/workspaces/use-workspace-surface-lookup";
 import { useSessionDirectoryStore } from "@/stores/sessions/session-directory-store";
 import { getSessionRecord } from "@/stores/sessions/session-records";
